# docs/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Pebble AI Backend API
  description: |
    Pebble AI 后端 API 文档。

    主要功能：
    - 读心翻译器：分析文本中的情感操控模式
    - 练习本：保存和管理练习条目
    - 模拟陪练场：情景模拟对话练习

    所有 API 响应遵循统一格式：
    - 成功: `{ "success": true, "data": {...} }`
    - 失败: `{ "success": false, "error": { "code": "...", "message": "...", "requestId": "..." } }`
  version: 1.0.0
  contact:
    name: Pebble AI Team
servers:
  - url: http://localhost:3000/api
    description: 本地开发环境
paths:
  /decode:
    post:
      summary: 分析文本情感
      description: |
        分析用户输入的文本，识别其中的情感操控模式，
        并提供回复建议。
      tags:
        - 读心翻译器
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecodeRequest'
      responses:
        '200':
          description: 分析成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 请求过于频繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /simulator:
    post:
      summary: 模拟陪练场操作
      description: |
        处理陪练场会话的各种操作：
        - start: 开始新会话
        - continue: 继续现有会话
        - restart: 重启会话
      tags:
        - 模拟陪练场
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulatorRequest'
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 会话不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
            - requestId
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - RATE_LIMITED
                - TIMEOUT
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
            message:
              type: string
            details:
              type: object
            requestId:
              type: string

    # Decode API Schemas
    DecodeRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 5000
          description: 需要分析的文本
          example: "你总是这样自私"
        context:
          type: string
          maxLength: 1000
          description: 可选的上下文信息
...
    SimulatorRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [start, continue, restart]
        scenarioId:
          type: string
          description: 开始新会话时必需
        sessionId:
          type: string
          format: uuid
          description: 继续或重启会话时必需
        message:
          type: string
          description: 继续会话时必需
...